+++ zfs/replicate/task/execute.py
@@ -3,7 +3,7 @@
 import itertools
 from typing import List, Tuple
 
-from .. import filesystem, optional, snapshot
+from .. import filesystem, optional, receive, send, snapshot
 from ..compress import Compression
 from ..filesystem import FileSystem
 from .type import Action, Task
@@ -13,9 +13,9 @@ def execute(  # pylint: disable=R0917,R0913
     remote: FileSystem,
     tasks: List[Tuple[FileSystem, List[Task]]],
     ssh_command: str,
-    follow_delete: bool,
     compression: Compression,
-    raw: bool,
+    send_options: send.Options,
+    receive_options: receive.Options,
 ) -> None:
     """Execute all tasks."""
     sorted_tasks = sorted(tasks, key=lambda x: len(x[0].name.split("/")), reverse=True)
@@ -38,9 +38,9 @@ def execute(  # pylint: disable=R0917,R0913
                     remote,
                     a_tasks,
                     ssh_command=ssh_command,
-                    follow_delete=follow_delete,
                     compression=compression,
-                    raw=raw,
+                    send_options=send_options,
+                    receive_options=receive_options,
                 )
 
 
@@ -61,9 +61,9 @@ def _send(  # pylint: disable=R0917,R0913
     remote: FileSystem,
     tasks: List[Task],
     ssh_command: str,
-    follow_delete: bool,
     compression: Compression,
-    raw: bool,
+    send_options: send.Options,
+    receive_options: receive.Options,
 ) -> None:
     for task in tasks:
         snapshot.send(
@@ -71,7 +71,7 @@ def _send(  # pylint: disable=R0917,R0913
             optional.value(task.snapshot),
             ssh_command=ssh_command,
             compression=compression,
-            follow_delete=follow_delete,
-            raw=raw,
+            send_options=send_options,
+            receive_options=receive_options,
             previous=optional.value(task.snapshot).previous,
         )