+++ zfs_test/replicate_test/send_test/__init__.py
@@ -1,0 +1,1 @@
+"""Test send functionality."""